typedesc: remove HydrateTypeInfoWithName, simplify EnsureTypeIsHydrated
This refactor simplifies the hydration logic owned by type descriptors
without changing it in any significant way.

Release note: None
